% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeCountsFromVector.R, R/hello.R
\name{aeCountsFromVector}
\alias{aeCountsFromVector}
\title{aeCountsFromVector;
Function to convert a vector into a matrix with rows corresponing to the subjects and collums to the PTs}
\usage{
aeCountsFromVector(pts)

aeCountsFromVector(pts)
}
\arguments{
\item{pts}{vector with pts ID, rownames are the subject IDs}
}
\value{
A matrix with rows corresponding to the subject, and collumns to the PTs

A matrix with rows corresponding to the subject, and collumns to the PTs
}
\description{
aeCountsFromVector;
Function to convert a vector into a matrix with rows corresponing to the subjects and collums to the PTs

aeCountsFromVector;
Function to convert a vector into a matrix with rows corresponing to the subjects and collums to the PTs
}
