% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAHAMADataSet.R
\name{BAHAMADataSet}
\alias{BAHAMADataSet}
\title{BAHAMADataSet object and constructors}
\usage{
BAHAMADataSet(
  aeCounts,
  medDRAData,
  sampleData,
  thresholdIncidence = 5,
  thresholdStructure = 2,
  alternative_weights = 0
)

BAHAMADataSet(
  aeCounts,
  medDRAData,
  sampleData,
  thresholdIncidence = 5,
  thresholdStructure = 2,
  alternative_weights = 0
)
}
\arguments{
\item{aeCounts}{A matrix of non-negative integers. Each column of aeCounts correspond to a single AE, each row to a sample.}

\item{sampleData}{A vector or matrix, with at least 1 column with the treatment group. Each row corresponds to a single sample.}

\item{thresholdIncidence}{A non-negative interger with the minimal incidence of a single AE}

\item{thresholdStructure}{A non-negative interger with the minimal number of siblings of a single AE}

\item{alternative_weights}{A list with a PT weight matrix, a HLT weight matrix and a HLGT weight matrix}

\item{BAHAMAData}{A medDraData object}
}
\value{
A object to store the input values for the BhaMae method.
}
\description{
BAHAMADataSet object and constructors

BAHAMADataSet object and constructors
}
\section{Slots}{

\describe{
\item{\code{N_trt}}{The total number of patients in the treatment group}

\item{\code{N_cont}}{The total number of patients in the control group}

\item{\code{N_ae}}{The total number of adverse events, after data aggregation}

\item{\code{N_pt}}{The total number of PTs, after data aggregation}

\item{\code{N_hlt}}{The total number of HLTs, after data aggregation}

\item{\code{N_hlt_noPT}}{The total number of HLTs, without a corresponing PT, after data aggregation}

\item{\code{N_hlt_PT}}{The total number of HLTs, with a corresponing PT, after data aggregation}

\item{\code{N_hlgt}}{The total number of HLGTs, after data aggregation}

\item{\code{N_hlgt_noHLT}}{The total number of HLGTs, without a corresponing HLT, after data aggregation}

\item{\code{N_hlgt_hlt}}{The total number of HLGTs, with a corresponing HLT, after data aggregation}

\item{\code{N_soc}}{The total number of SOCs, after data aggregation}

\item{\code{y_hlgt_trt}}{Per HLGTs the incidence in the treatment group, after data aggregation}

\item{\code{y_hlgt_cont}}{Per HLGTs the incidence in the control group, after data aggregation}

\item{\code{y_hlt_trt}}{Per HLTs the incidence in the treatment group, after data aggregation}

\item{\code{y_hlt_cont}}{Per HLTs the incidence in the control group, after data aggregation}

\item{\code{y_pt_trt}}{Per PTs the incidence in the treatment group, after data aggregation}

\item{\code{y_pt_cont}}{Per PTs the incidence in the control group, after data aggregation}

\item{\code{SOC_HLGT}}{The mapping between the HLGTs and SOCs}

\item{\code{HLGT_HLT}}{The mapping between the HLTs and HLGTs}

\item{\code{HLT_PT}}{The mapping between the PTs and HLTs}

\item{\code{N_PT_perHLT}}{The number of PTs per HLT, to correct for data aggregation}

\item{\code{N_PT_perHLGT}}{The number of PTs per HLGT, to correct for data aggregation}
}}

\examples{
add(1, 1)
add(10, 1)
}
