% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medDRAData.R
\docType{class}
\name{medDRAData-class}
\alias{medDRAData-class}
\alias{medDRAData}
\title{medDRAData object and constructors}
\description{
medDRAData object and constructors
}
\section{Slots}{

\describe{
\item{\code{pt_hlt}}{Mapping between the PTs and HLTs}

\item{\code{hlt_hlgt}}{Mapping between the HLTs and HLGTs}

\item{\code{hlgt_soc}}{Mapping between the HLGTs and SOCs}
}}

