% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medDRAData.R
\name{medDRADataFromMatrix}
\alias{medDRADataFromMatrix}
\title{medDRAData object and constructors}
\usage{
medDRADataFromMatrix(medDRAMatrix)
}
\arguments{
\item{medDRAMatrix}{A matrix of MedDRA IDs. The columns correspond to the MedDRA structure (SOC, HLGT, HLT, PT), each row correspond to a single AE. If multiaxiallity is present, a single AE is represented in multiple rows.}
}
\description{
medDRAData object and constructors
}
